Passage de EqLogic avec les modeles
